% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMS.computeProfit.R
\name{BMS.computeProfit}
\alias{BMS.computeProfit}
\title{Compute profit for a BMS Pricer study}
\usage{
BMS.computeProfit(
  utils,
  prices,
  simPrices,
  simSKUs = NULL,
  nlev,
  weight = NULL,
  none = FALSE,
  iaw = NULL,
  FC = FALSE,
  costs
)
}
\arguments{
\item{utils}{A matrix containing the utilities}

\item{prices}{A list containing the prices used in the interview}

\item{simPrices}{A vector containing the prices to be used in the simulated scenario}

\item{simSKUs}{A vector containing indicies of the SKUs to be used in the simulated scenario}

\item{nlev}{An vector indicating the number of levels per attribute}

\item{weight}{A vector with the weights (one per respondent). Will be set to 1 if \code{NULL}}

\item{none}{A boolean variable indicating if NONE should be used in the simulated scenario. - default \code{FALSE}}

\item{iaw}{A matrix of individual awareness factors corresponding to those in the BMS Pricer tool.
All are set to 1 if \code{NULL}. In case both individual awareness and distribution factors need to be used
the respective matrices need to be multiplied before passing to this function.}

\item{FC}{A boolean variable indicating if first choice simulation should be used (\code{FALSE} indicates
preference share simulation) - default \code{TRUE}}

\item{costs}{A vector indicating the costs per SKU in the model.}
}
\value{
A list including elements
  \item{profit}{mean profit across all respondents}
  \item{demand}{mean share of the scenario across all respondents}
}
\description{
Calculates the profit for a BMS Pricer like study.
}
\examples{

beer_def <- beer_data$def

profit_Beer <- VP.computeProfit(beer_data$utils_mat,
                                beer_def$prices,
                                beer_def$prices[,3],
                                simSKUs = NULL,
                                nlev = beer_data$nlev,
                                weight = beer_data$weight,
                                none = FALSE,
                                iaw = NULL,
                                FC = FALSE,
                                costs = 0.6 * beer_def$prices[,3])

profit_Beer$profit

}
\author{
Maximilian Rausch - m.rauschh@bms-net.de
}
