% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMS.read_batchfile.R
\name{BMS.read_batchfile}
\alias{BMS.read_batchfile}
\title{Read BMS Pricer definitions file}
\usage{
BMS.read_batchfile(file)
}
\arguments{
\item{file}{A string value with the path to the XLSM batch file file to import.}
}
\value{
A list including elements
  \item{scenarios}{A data.table including the scenarios}
  \item{prices}{A list including the prices per SKU}
  \item{clusters}{A data.table including the clusters defined in the batch file}
  \item{segments}{A list with the defined segments}
  \item{info}{A list with infos: Scenario labels, Clustering, Segment, Weighted}
}
\description{
Reads the BMS Pricer def file extracting the labels and prices
}
\examples{

\dontrun{
BMS.read_batchfile(file = "data/TEST_FILE.def")
}

}
\author{
Maximilian Rausch - Maximilian.Rausch@bms-net.de
}
