% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMS.read_def.R
\name{BMS.read_def}
\alias{BMS.read_def}
\title{Read BMS Pricer definitions file}
\usage{
BMS.read_def(file)
}
\arguments{
\item{file}{A string value with the path to the DEF file to import.}
}
\value{
A list including elements
  \item{brands}{A vector including the SKU labels}
  \item{prices}{A list including the prices per SKU}
  \item{segDef}{A list including the segment labels}
}
\description{
Reads the BMS Pricer def file extracting the labels and prices
}
\examples{

\dontrun{
VP.read_def(file = "data/TEST_FILE.def")
}

}
\author{
Maximilian Rausch - Maximilian.Rausch@bms-net.de
}
