% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibEXE.R
\name{calibEXE}
\alias{calibEXE}
\title{Calib EXE}
\usage{
calibEXE(
  BWconcepts = NULL,
  PI = NULL,
  utils = NULL,
  cut = 42,
  nlev = NULL,
  PIsteps = c(0.95, 0.5, 0.3, 0.15, 0.05)
)
}
\arguments{
\item{BWconcepts}{A matrix/data.frame containing the best and worst concepts per respondents.}

\item{PI}{A matrix containing the answer to the purchase intention questions
for the best/worst concepts.}

\item{utils}{A matrix containing the utilities per respondent of the model.}

\item{cut}{An integer value indicating the maximal value for utilities after calibration.}

\item{nlev}{A vector indicating the number of levels per attribute}

\item{PIsteps}{A vector with the purchse intention %-value to be used for the calibration
(default: \code{c(.95, .5, .3, .15, .05)})}
}
\value{
A list including 13 elements
\item{BWconcepts}{A matrix/data.frame containing the best and worst concepts per respondents
which have been passed to the function}
\item{utils_calib}{A matrix containing the CALIBRATED utilities per respondent of the model.}
\item{utils}{A matrix containing the original utilities per respondent of the model.}
\item{check_order_BW}{A vector containing boolean values indicating if the utilities
for the best concept have been greater than the ones of the worst concept.}
\item{check_order_PI}{A vector containing boolean values indicating if the purchase intentions
for the best concept have been greater than the ones of the worst concept.}
\item{utl_sum}{A matrix containing the corrected utility sums for the best and worst concept.
Values in the wrong order are set to be equal.}
\item{PurchaseInt}{A matrix containing the corrected purchase intentions for the best
and worst concept. (1 = .95, 2 = .5, 3 = .3, 4 = .15, 5 = .05)
Values in the wrong order are set to be equal.}
\item{utl_sum_ORIG}{A matrix containing the original (uncorrected) utility sums for
the best and worst concept.}
\item{PurchaseInt_ORIG}{A matrix containing the original (uncorrected) purchase intentions for the best
and worst concept.}
\item{a}{The slope of the linear function used for the calibration.}
\item{b}{The intercept of the linear function used for the calibration.}
\item{nlev}{A vector indicating the number of levels per attribute}
\item{cut}{An integer indicating the cutoff value used for the maximal
value for utilities after calibration.}
}
\description{
Calibrates the utilities based on purchase intention questions (calibEXE)
}
\examples{

\dontrun{
calibData <- calibEXE(BWconcepts = BWconcepts[, c(paste0("B_Att_", sequence(natt)),
                                                  paste0("W_Att_", sequence(natt)))],
PI = BWconcepts[, c("PI_B", "PI_W")],
utils = dat_input$utils_mat,
cut = 42, nlev = nlev)
}

}
\author{
Maximilian Rausch - Maximilian.Rausch@tns-infratest.com
}
