% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConceptOpt_ISBC.R
\name{ConceptOpt_ISBC}
\alias{ConceptOpt_ISBC}
\title{Concept optimization with ISBC}
\usage{
ConceptOpt_ISBC(
  dat = NULL,
  def = NULL,
  FrameData = NULL,
  calib = TRUE,
  nlev = NULL,
  cut = 42,
  ID = "Respondent_Serial",
  PIsteps = c(0.95, 0.5, 0.3, 0.15, 0.05)
)
}
\arguments{
\item{dat}{A string indicating the path/filename of the ValueDriver dat-file.}

\item{def}{A string indicating the path/filename of the ValueDriver def-file.}

\item{FrameData}{A string indicating the path/filename of the frame questionnaire data.}

\item{calib}{An boolean value indicating if the utilities should be calibrated on purchase
intention - default = \code{TRUE}.}

\item{nlev}{A vector indicating the number of levels per attribute}

\item{cut}{An integer value indicating the maximal value for utilities after calibration.
(required for \code{\link{calibEXE}})}

\item{ID}{A string indicating the name of teh ID variable in the SPSS dataset - default = \code{"Respondent_Serial"}.}

\item{PIsteps}{A vector with the purchse intention %-value to be used for the calibration
(default: \code{c(.95, .5, .3, .15, .05)})}
}
\value{
A list including 13 elements
\item{call}{A string returning the call including the inputs}
\item{calibData}{A list with the values of the \code{\link{calibEXE}} call.
(\code{NULL} if \code{calib == FALSE}))}
\item{excel_export}{A data.frame containing the export matrix;
concept definition + aggregated shares + sd + individual shares.}
\item{excel_export_calib}{A data.frame containing the calibrated export matrix;
concept definition + aggregated shares + sd + individual shares. (\code{NULL} if \code{calib == FALSE}))}
\item{dat_file_calib}{A data.frame containing the export information for the
calibrated ValueDriver dat-file. (\code{NULL} if \code{calib == FALSE}))}
\item{overview}{A data.frame with additional information on the calibExe step.
(\code{NULL} if \code{calib == FALSE}))}
\item{path}{A string returning the working directory where the export files had been saved to.}
\item{weight}{A vector with the weight used for aggregation}
}
\description{
Runs the process for "concept optimization with ISBC" studies - Always weighted!
}
\examples{

\dontrun{
nlev <- c(4, 4, 4, 2, 4, 2, 4, 6)
ConcOptTest <- ConceptOpt_ISBC("Innogy.dat",
                               "Innogy.def",
                               FrameData = "./data_calib.SAV",
                               calib = TRUE,
                               nlev = nlev,
                               cut = 42,
                               ID = "Respondent_Serial")
str(ConcOptTest)
}

}
\author{
Maximilian Rausch - Maximilian.Rausch@tns-infratest.com
}
