% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolPrice.R
\name{interpolPrice}
\alias{interpolPrice}
\title{Design matrix with interpolated price}
\usage{
interpolPrice(
  scenario,
  priceSim,
  prices,
  scenDUMMY,
  priceInd,
  priceInd_Dummy,
  nlev,
  scenInd = 1,
  brandAtt = 1
)
}
\arguments{
\item{scenario}{A matrix with the scenario for which the Strategy Profile should be calculated}

\item{priceSim}{the price to be simulated}

\item{prices}{A list containing the prices for the brand specific price attributes,
e.g. the part from from \code{\link{VD.read_def}}}

\item{scenDUMMY}{A matrix with the dummy coded scenario}

\item{priceInd}{A vector with the index values for the brand specific price attributes}

\item{priceInd_Dummy}{A vector with the index values for the brand specific price attributes
in the dummy coded scenario}

\item{nlev}{A vector with the number of levels per attribute}

\item{scenInd}{An integer that indicates the scenario for which the Strategy Profile should be calculated}

\item{brandAtt}{An integer indicating the brand attribute in case of brand specific prices}
}
\value{
A matrix with the dummy coded design matrix including the interpolation for the price.
}
\description{
Include a interpolated brand specific price in a design matrix
}
\examples{

\dontrun{
# Scenario
utils_mat <- VMinR$VDdata$utils_mat
defIN <- VD.read_def(file = "data/TEST_timtim_5seg_2_gew_gew2.def",
                     nlev = c(4, 6, 4, 2, 2, 2, 2, 2, 2, 2, 2, 5, 5, 5, 5, 5, 5))
nlev <- defIN$nlev

scenario <- rbind(c(1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0),
                  c(1, 2, 1, 1, 1, 1, 1, 1, 1, 1, 1, 0, 1, 0, 0, 0, 0, 0),
                  c(1, 3, 1, 1, 1, 1, 1, 1, 1, 1, 1, 0, 0, 1, 0, 0, 0, 0),
                  c(1, 4, 1, 1, 1, 1, 1, 1, 1, 1, 1, 0, 0, 0, 1, 0, 0, 0),
                  c(1, 5, 1, 1, 1, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 1, 0, 0),
                  c(1, 6, 1, 1, 1, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 1, 0),
                  c(0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1))

priceInd <- grep("^Price", names(defIN$att_List))

prices <- lapply(defIN$att_List[priceInd],
function(p) {
  as.numeric(gsub("€ ", "", p))
})

scenDUMMY <- convertSSItoDesign(scenario,
                                nlev = c(nlev, 1))
priceInd_Dummy <- grep(paste0("ATT(", paste0(priceInd, collapse = "|"), ")"), names(scenDUMMY))

scenINTER <- interpolPrice(priceSim = 300,
                           prices = prices,
                           scenDUMMY = scenDUMMY,
                           priceInd_Dummy = priceInd_Dummy,
                           nlev = nlev,
                           scenInd = 1,
                           brandAtt = 2)

scenINTER
}

}
\author{
Maximilian Rausch - Maximilian.Rausch@bms-net.de
}
