% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nashEquilibrium.R
\name{nashEquilibrium}
\alias{nashEquilibrium}
\title{Compute optimal prices for a ValuePricer study}
\usage{
nashEquilibrium(
  utils,
  simSKUs,
  startingPrices,
  prices,
  pr_range_mat = NULL,
  nlev,
  costs,
  grp,
  line_price,
  freeze = NULL,
  FC = FALSE,
  none = FALSE,
  weight = NULL,
  iaw = NULL,
  constraints = NULL,
  optKPI = "profit",
  shiny = FALSE
)
}
\arguments{
\item{utils}{A matrix containing the utilities}

\item{simSKUs}{A vector containing indicies of the SKUs to be used in the simulated scenario}

\item{startingPrices}{A vector containing the prices to be used as starting point for the optimisation}

\item{prices}{A matrix containing the prices used in the interview}

\item{pr_range_mat}{A matrix (2 columns) containing the upper and lower boundaries for the optimisation.}

\item{nlev}{A vector indicating the number of levels per attribute}

\item{costs}{A vector indicating the costs per SKU in the model.}

\item{grp}{A vector indicating which SKUs should be considered as a group
(e.g. from the same brand) during optimisation. Same index indicates a group.
1st column: sequencial index.
2nd column: indicator for groups.
3rd column: indicator for line pricing}

\item{line_price}{A boolean variable telling if line pricing should be applied}

\item{freeze}{A vector inculding the SKU numbers that should be excluded from the optimisation.
default = \code{NULL}}

\item{FC}{A boolean variable indicating if first choice simulation should be used (\code{FALSE} indicates preference
share simulation) - default \code{FALSE}}

\item{none}{A boolean variable indicating if NONE should be used in the simulated scenario.
default \code{FALSE}}

\item{weight}{A vector with the weights (one per respondent). Will be set to 1 if \code{NULL}}

\item{iaw}{A matrix of individual awareness factors corresponding to those in the ValuePricer tool.
All are set to 1 if \code{NULL}. In case both individual awareness and distribution factors need to be used
the respective matrices need to be multiplied before passing to this function.}

\item{constraints}{A vector indicating the allowed increase/decrease in share, profit or revenue
(in percentage points). If \code{NULL} (default) no constraints will be applied.}

\item{optKPI}{A string indicating the KPI to be optimized (profit, revenue, share). default \code{"profit"}.}

\item{shiny}{A boolean variable stating if the function is used in the shiny app. Adds progress bar parts.
default \code{TRUE}.}
}
\value{
A vector of optimal prices given the specs.
}
\description{
Calculates the optimal prices (based on the theory of John Nash) for a ValuePricer like study.
}
\examples{

\dontrun{
beer_def <- beer_data$def
design_beer <- diag(length(beer_def$brands))

input_list_Nash <- list(utils = beer_data$utils_mat[, -151],
                        simSKUs = seq_along(length(beer_def$brands)),
                        startingPrices = beer_def$prices[,3],
                        prices = beer_data$pricemat_tested,
                        pr_range_mat = NULL,
                        nlev = beer_data$nlev,
                        costs = beer_data$pricemat_tested[,1] * 0.8,
                        grp = NULL,
                        line_price = FALSE,
                        freeze = NULL,
                        FC = FALSE,
                        none = FALSE,
                        weight = beer_data$weight,
                        iaw = NULL,
                        constraints = NULL,
                        optKPI = "profit")

# Nash optimization
Nash_run <- do.call(nashEquilibrium, input_list_Nash)
}

}
\author{
Maximilian Rausch - m.rausch@bms-net.de
}
