% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProductAcceptance.R
\name{prodAcceptance}
\alias{prodAcceptance}
\title{Calculate Product Acceptance (Buying Rate)}
\usage{
prodAcceptance(conc, utils)
}
\arguments{
\item{conc}{A binary vector containing 0s or 1s for each level of the model.}

\item{utils}{A matrix containing the utilities per respondent of the model.
Remove the NONE column if present.}
}
\value{
A vector contaning the prodct acceptance/buying rate values of the concept for each respondent.
}
\description{
Calculates the Productacceptance/Buying Rate for a given concept.
}
\examples{

\dontrun{
...
}

}
\author{
Maximilian Rausch - Maximilian.Rausch@tns-infratest.com
}
