% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simWTP.R
\name{simWTP}
\alias{simWTP}
\title{Simulate scenario with interpolated price}
\usage{
simWTP(
  pr,
  prices = prices,
  scenDUMMY_WTP,
  scenario_WTP,
  priceInd,
  priceInd_Dummy,
  nlev,
  scenInd = 1,
  brandAtt = 1,
  utils,
  nlevIN = c(nlev, 1),
  weight = NULL,
  FC = FALSE
)
}
\arguments{
\item{pr}{price level to be interpolated (needs to be with the range of prices)}

\item{prices}{A list containing the prices for the brand specific price attributes,
e.g. the part from from \code{\link{VD.read_def}}}

\item{scenDUMMY_WTP}{A matrix with the dummy coded scenario}

\item{scenario_WTP}{A matrix with the scenario}

\item{priceInd}{A vector with the index values for the brand specific price attributes}

\item{priceInd_Dummy}{A vector with the index values for the brand specific price attributes
in the dummy coded scenario}

\item{nlev}{A vector with the number of levels per attribute}

\item{scenInd}{An integer that indicates the scenario for which the Strategy Profile should be calculated}

\item{brandAtt}{An integer indicating the brand attribute in case of brand specific prices.
\code{NULL} in case of no brand specific price.}

\item{utils}{A matrix containing the utilities per respondent of the model.}

\item{nlevIN}{A vector with the number of levels per attribute including the \code{NONE} if applicable}

\item{weight}{A vector with the weights (one per respondent)}

\item{FC}{A boolean variable indicating if first choice simulation should be
used (\code{FALSE} indicates preference share simulation) - default \code{FALSE}}
}
\value{
A numeric value returning the preference share for the indicated scenario
}
\description{
Simulate scenario with interpolated price to be able to calculate simulation based WTP
}
\examples{

\dontrun{
simWTP(...)
}

}
\author{
Maximilian Rausch - Maximilian.Rausch@bms-net.de
}
