% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StrategyProfile.R
\name{StrategyProfile}
\alias{StrategyProfile}
\title{Compute Strategy Profile}
\usage{
StrategyProfile(
  scenario,
  att_List,
  utils_mat,
  scenInd = 1,
  brandAtt = 1,
  priceInd = NULL,
  priceDefault = NULL
)
}
\arguments{
\item{scenario}{A matrix with the scenario for which the Strategy Profile should be calculated}

\item{att_List}{A list containing the attributes and levels, e.g. from \code{\link{VD.read_def}}}

\item{utils_mat}{A matrix containing the utilities}

\item{scenInd}{An integer that indicates the scenario for which the Strategy Profile should be calculated}

\item{brandAtt}{An integer indicating the brand attribute in case of brand specific prices}

\item{priceInd}{A vector with the index values for the brand specific price attributes.
If \code{NULL} the respective prices will be determined from the \code{att_List}}

\item{priceDefault}{A vector with the default price levels for each brand specific price attribute.
Default value is first level for all if \code{NULL}.}
}
\value{
A list including elements
  \item{Sim_Scen}{results of \code{\link{VD.computeShares}}} for \code{scenario}
  \item{Sim_DT}{data.table with mean share of choice}
  \item{StrategyAbs}{share of choice for the Strategy Profile results}
  \item{StrategyRel}{share of choice difference of the Strategy Profile results to \code{scenario} results}
  \item{Strategy_DT}{data.table with Strategy Profile (absolute and delta) for simulated concept}
}
\description{
Calculates the Strategy Profile for a Feature Conjoint study
}
\examples{

\dontrun{
# Scenario
utils_mat <- VMinR$VDdata$utils_mat
defIN <- VD.read_def(file = "data/TEST_timtim_5seg_2_gew_gew2.def",
                     nlev = c(4, 6, 4, 2, 2, 2, 2, 2, 2, 2, 2, 5, 5, 5, 5, 5, 5))

scenario <- rbind(c(1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0),
                  c(1, 2, 1, 1, 1, 1, 1, 1, 1, 1, 1, 0, 1, 0, 0, 0, 0, 0),
                  c(1, 3, 1, 1, 1, 1, 1, 1, 1, 1, 1, 0, 0, 1, 0, 0, 0, 0),
                  c(1, 4, 1, 1, 1, 1, 1, 1, 1, 1, 1, 0, 0, 0, 1, 0, 0, 0),
                  c(1, 5, 1, 1, 1, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 1, 0, 0),
                  c(1, 6, 1, 1, 1, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 1, 0),
                  c(0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1))

priceInd <- grep("^Price", names(defIN$att_List))
SP_Test <- StrategyProfile(
  scenario = scenario,
  att_List = defIN$att_List,
  utils_mat = utils_mat,
  scenInd = 1,
  brandAtt = 2,
  priceInd = grep("^Price", names(defIN$att_List)),
  priceDefault = rep(1, length(priceInd)))

SP_Test
}

}
\author{
Maximilian Rausch - Maximilian.Rausch@bms-net.de
}
