% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_calibEXE.R
\name{summary_calibEXE}
\alias{summary_calibEXE}
\title{Summary of calibEXE}
\usage{
summary_calibEXE(
  calibData = calibData,
  outfile = "summary.txt",
  infile = "INFILE"
)
}
\arguments{
\item{calibData}{A list containing the results of \code{\link{calibEXE}}}

\item{outfile}{A text value indicating the label of the output file.
Default \code{"summary.txt"}.}

\item{infile}{A text value indicating the label input file for
\code{\link{calibEXE}} function. Default \code{"INFILE"}.}
}
\value{
No output generated. Summary will be exported to the indicated file.
}
\description{
Runs a summary for the calibration of utilities using the calibEXE function.
}
\examples{

\dontrun{
summary_calibEXE(calibData, "summary.txt", infile = dat_input$def$file_in)
}

}
\author{
Maximilian Rausch - Maximilian.Rausch@tns-infratest.com
}
