% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VP.computeShares.R
\name{VP.computeShares}
\alias{VP.computeShares}
\title{Compute ValuePricer shares}
\usage{
VP.computeShares(
  utils,
  prices,
  simPrices,
  simSKUs = NULL,
  nlev,
  weight = NULL,
  none = FALSE,
  iaw = NULL,
  FC = FALSE
)
}
\arguments{
\item{utils}{A matrix containing the utilities}

\item{prices}{A matrix containing the prices used in the interview}

\item{simPrices}{A vector containing the prices to be used in the simulated
scenario}

\item{simSKUs}{A vector containing indicies of the SKUs to be used in the
simulated scenario}

\item{nlev}{An vector indicating the number of levels per attribute}

\item{weight}{A vector with the weights (one per respondent). Will be set to
1 if \code{NULL}}

\item{none}{A boolean variable indicating if NONE should be used in the
simulated scenario. - default \code{FALSE}}

\item{iaw}{A matrix of individual awareness factors corresponding to those
in the ValuePricer tool. All are set to 1 if \code{NULL}. In case both
individual awareness and distribution factors need to be used the respective
matrices need to be multiplied before passing to this function.}

\item{FC}{A boolean variable indicating if first choice simulation should be
used (\code{FALSE} indicates preference share simulation) - default
\code{TRUE}}
}
\value{
A list including elements \item{ind_sim}{individual shares for each
respondent} \item{Xbeta}{\code{X * beta} matrix used to calculate the shares
(utility sums).} \item{simPrices}{respective input object passed through}
\item{simSKUs}{respective input object passed through}
\item{prices}{respective input object passed through} \item{iaw}{respective
input object passed through} \item{none}{respective input object passed
through} \item{weight}{respective input object passed through}
\item{simShares}{aggregated shares accross all respondents}
}
\description{
Calculates the shares (first choice or preference share) for a ValuePricer
like study.
}
\examples{

beer_def <- beer_data$def

sim_Beer <- VP.computeShares(beer_data$utils_mat,
                             beer_def$prices,
                             beer_def$prices[,3],
                             simSKUs = NULL,
                             nlev = beer_data$nlev,
                             weight = beer_data$weight,
                             none = FALSE,
                             iaw = NULL,
                             FC = FALSE)

round(sim_Beer$simShares, 3)

}
\author{
Maximilian Rausch - Maximilian.Rausch@tns-infratest.com
}
