% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WTP.R
\name{WTP}
\alias{WTP}
\title{Simulation based WTP}
\usage{
WTP(
  scen,
  scenIndex,
  nLevels,
  attNames,
  brandAttribute,
  priceIndex,
  priceDefault,
  prices,
  utilsIN,
  none = TRUE,
  weight = NULL,
  FC = FALSE
)
}
\arguments{
\item{scen}{A matrix with the scenario}

\item{scenIndex}{An integer that indicates the concept for which the WTP should be calculated}

\item{nLevels}{A vector with the number of levels per attribute}

\item{attNames}{A list with attribute and level labels (e.g. \code{att_list} from \code{\link{VD.read_def}})}

\item{brandAttribute}{An integer indicating the brand attribute in case of brand specific prices.
\code{NULL} in case of no brand specific price.}

\item{priceIndex}{A vector with the index values for the brand specific price attributes}

\item{priceDefault}{A vector with the default levels to be used for prices (in case of asd)}

\item{prices}{A list containing the prices for the brand specific price attributes,
e.g. the part from from \code{\link{VD.read_def}}}

\item{utilsIN}{A matrix containing the utilities per respondent of the model.}

\item{none}{A boolean variable indicating if NONE should be used in the simulated scenario. - default \code{TRUE}}

\item{weight}{A vector with the weights (one per respondent)}

\item{FC}{A boolean variable indicating if first choice simulation should be
used (\code{FALSE} indicates preference share simulation) - default \code{FALSE}}
}
\value{
A list including elements
  \item{WTP_List}{A list containing the WTP values per attribute and level}
  \item{WTP_Scen}{A list containing the scenarios used for WTP calculation}
  \item{ref_Scen}{A list with the simulation results for the reference scenario (\code{\link{VD.computeShares}})}
  \item{WTP_DT}{A data.table with the WTP values for each level (excl. price attributes)}
}
\description{
Function used to calculate simulation based WTP for all attribute levels of a scenario
}
\examples{

\dontrun{
WTP(...)
}

}
\author{
Maximilian Rausch - Maximilian.Rausch@bms-net.de
}
