% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StrategyProfile.R
\name{StrategyProfile}
\alias{StrategyProfile}
\title{Compute Strategy Profile}
\usage{
StrategyProfile(
  scenario,
  att_List,
  utils_mat,
  scenInd = 1,
  asdRules = NULL,
  defaultLev = NULL,
  SimModel = "SoC"
)
}
\arguments{
\item{scenario}{A matrix with the scenario for which the Strategy Profile should be calculated}

\item{att_List}{A list containing the attributes and levels, e.g. from \code{\link{VD.read_def}}}

\item{utils_mat}{A matrix containing the utilities}

\item{scenInd}{An integer that indicates the scenario for which the Strategy Profile should be calculated}

\item{asdRules}{A list with asd rules
\describe{
  \item{baseAtt}{base attribute for asd rule}
  \item{baseLev}{level of base attribute}
  \item{asdAtt}{asd attributes. Index values of th eattributes that are excluded with th ebase attribute}
  }}

\item{defaultLev}{A integer or vector with same nength as number of asd attributes with the default level
for asd attributes attribute.
Default value is first level for all if \code{NULL}.}

\item{SimModel}{A character value indicating the Simulation Model.
\describe{
  \item{SoC}{Share of Choice - Preference Share (Default)}
  \item{FC}{First Choice}
  }}
}
\value{
A list including elements
  \item{Sim_Scen}{results of \code{\link{VD.computeShares}}} for \code{scenario}
  \item{Sim_DT}{data.table with mean share of choice}
  \item{StrategyAbs}{share of choice for the Strategy Profile results}
  \item{StrategyRel}{share of choice difference of the Strategy Profile results to \code{scenario} results}
  \item{Strategy_DT}{data.table with Strategy Profile (absolute and delta) for simulated concept}
}
\description{
Calculates the Strategy Profile for a Feature Conjoint study
}
\examples{

\dontrun{
# Scenario
utils_mat <- VMinR$VDdata$utils_mat
defIN <- VD.read_def(file = "data/TEST_timtim_5seg_2_gew_gew2.def",
                     nlev = c(4, 6, 4, 2, 2, 2, 2, 2, 2, 2, 2, 5, 5, 5, 5, 5, 5))

scenario <- rbind(c(1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0),
                  c(1, 2, 1, 1, 1, 1, 1, 1, 1, 1, 1, 0, 1, 0, 0, 0, 0, 0),
                  c(1, 3, 1, 1, 1, 1, 1, 1, 1, 1, 1, 0, 0, 1, 0, 0, 0, 0),
                  c(1, 4, 1, 1, 1, 1, 1, 1, 1, 1, 1, 0, 0, 0, 1, 0, 0, 0),
                  c(1, 5, 1, 1, 1, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 1, 0, 0),
                  c(1, 6, 1, 1, 1, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 1, 0),
                  c(0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1))

priceInd <- grep("^Price", names(defIN$att_List))
SP_Test <- StrategyProfile(
  scenario = scenario,
  att_List = defIN$att_List,
  utils_mat = utils_mat,
  scenInd = selConc,
  asdRules = asd_rules,
  defaultLev = selPrice,
  SimModel = "SoC")

SP_Test
}

}
\author{
Maximilian Rausch - Maximilian.Rausch@bms-net.de
}
