\name{VMinR}
\alias{VMinR}
\title{VM test data - timtim}
\format{A RData file containg the test data
\describe{
   \item{VDdata}{A list containing the output of \code{\link{get_DriverData}}}
   \item{basecase}{A matrix containing the scenario information in SSI style}
   \item{basecase_dummy_2}{A matrix containing the scenario information in dummy coding}
   \item{natt}{A variable indicating the number of attribute for the timtim test case}
   \item{nlev}{A vector indicating the number of number of levels per attribute for the timtim test case}
   \item{sim_BaseR}{A vector containing the preferecne shares for the basecase of the timtim test case.}
}}
\description{
Dataset containing the variable contents (data model) settings to perform input validation.
}
\usage{
VMinR
}
\keyword{datasets}
