% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VMinR-package.R
\docType{data}
\name{VMinR}
\alias{VMinR}
\title{VM test data - timtim}
\format{A list with 3 elements.  \describe{ \item{VDdata}{A list containing
the output of \code{\link{get_DriverData}}} \item{basecase}{A matrix
containing the scenario information in SSI style} \item{basecase_dummy_2}{A
matrix containing the scenario information in dummy coding} }}
\description{
Dataset containing the variable contents (data model) settings to perform
input validation.
}
\examples{

data(VMinR)
str(VMinR)

}
\keyword{datasets}
