% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VP.read_def.R
\name{VP.read_def}
\alias{VP.read_def}
\title{Read ValuePricer definitions file}
\usage{
VP.read_def(file)
}
\arguments{
\item{file}{A string value with the path to the DEF file to import.}
}
\value{
A list including elements \item{brands}{A vector including the SKU
labels} \item{prices}{A matrix (nSKUs x nPrices) including the prices per
SKU}
}
\description{
Reads the ValuePricer def file extracting the labels and prices
}
\examples{

\dontrun{
VP.read_def(file = "data/TEST_FILE.def")
}

}
\author{
Maximilian Rausch - Maximilian.Rausch@tns-infratest.com
}
